将player cache 公开，方便预加载处理
